[devops] Fix provisioning device testing bots.
It seems we only want to provision a few specific things, so modify the
selection logic to first ignore everything, then enable what we want.

This way we don't have to add new ignore flags here every time we add
something new to the provisioning script.

We also don't need to check if 7z and python3 provisioning is possible, since
we already know that (by manually checking the system-dependencies.sh script
for the current hash).