Add separate version variables for nugets.
Add separate version variables for our nugets. Each nuget will be version
according to its target platform, which means that we can re-use the version
for iOS (because we've matched iOS versions with our Xamarin.iOS packages),
but we need a separate version for watchOS and macOS, which do not match their
underlying platforms. And then I added a separate version variable for tvOS as
well, to make things match up nicely.