[Harness] Make sure we do not copy/generate the app sources more than once.
Threading is hard, even with async. We reached a situation in which the
src code of the test applications from the template was being generated
more thanonce. Ideally, we could have a nice solution with
AsyncLazy<bool> and use the lazy async to copy the code, but it was
moved to netcore 5. The fix is a little uglier but valid, lock and
enure we only generate it once. Please note that you CANNOT use a lock
statement inside a async method.

Fixes: https://github.com/xamarin/xamarin-macios/issues/8240