[Harness] Create IBuildToolTask and use it in the RunTestTask.
In order to de-couple the RunTestTask from Jenkins, create an interface
to be implemented, which is pass to use as a member (Composition
pattern). In order to do that, do not expose Jenkins as a property of
the interface because it is required just by the base constructor.

Moving the the use an interfance meant a lot of small changes that
should have no real effect (the compiler should have caught any possible
issues).