[msbuild] Use shared projects to avoid ILRepack.
Purpose: to get debug symbols (and file name / line number information in stack traces).

Caveat: this needs major work to ensure it works with windows. Also
confirmation that having more assemblies doesn't end up with assembly version
conflicts. Which it probably will, so this commit is most likely a no-go for
master.