Use a GCHandle instead of WeakReference in our IntPtr->NSObject table.
A WeakReference contains a GCHandle and provides additional services on top,
but we don't need those services, so we can just use a GCHandle instead.

This should decrease memory usage somewhat, since we won't have the
WeakReference objects around.

This also prepares us for switching the native API to use GCHandle instead of
passing MonoObject* to managed code (we won't have to re-create a GCHandle
inside a WeakReference when passed a GCHandle for a managed object).