Change signature to accept an `object`
This makes the call to `GetType` happens in the callee (once in IL)
instead of each individual callers.

```il
call instance class [mscorlib]System.Type [mscorlib]System.Object::GetType()
```

There's a single caller for now - but that will eventually change and the
balance will be negative (which is good for the total IL size).