[Harness] Add a new common build issue for Mac OS X test apps.
The issue is recurrent and happens when we try to build mac os x test
applications in which we cannot find the Xamarin.Mac.CSharp.targets. The
search string does not look for the error MSB4226 becuase that coudl be
due to other reasons, but we do know about the Xamarin.Mac targets.