[appkit] Allow ObjC calls into NSWindow.Close
The public `Close` API has additional logic before calling the internal
`_Close` method. However it's the latter that is decorated with the
`[Export]` attributes so it's what's called from ObjC.

This means that closing an NSWindow from the UI chrome (red button)
would by-pass the extra logic in `Close`, which make it possible to
`Dispose` the instance.

reference: https://github.com/xamarin/xamarin-macios/issues/8606