[generator] Add `ObjectDisposedException` when required on Handle access
Instead of providing `nil` to native API that are decorated **not**
accepting them.

Also do this using existing and new extension method helpers so the IL
size of the assemblies is (a bit) smaller than the existing code.

The latter has a small, positive impact for Xamarin.Mac (JIT'ing or
requiring IL for, non-full, AOT'ing) and for Xamarin.iOS when the
interpreter is used.