[foundation] Do not throw inside NSObject.SuperHandle property getter
It's against the FXDG https://docs.microsoft.com/en-us/dotnet/standard/design-guidelines/property

> ❌ AVOID throwing exceptions from property getters.

and it makes things harder in some corner cases, like when Apple calls
a selector after an instance is disposed, e.g.

https://github.com/xamarin/xamarin-macios/issues/8606#issuecomment-637221248

It's not the only case where we throw in getters but this one is an
immediate and much more common case than others.