[dotnet-linker] Add the MachO.cs file to get the Abi enum.
This also requires adding support for the ErrorHandler machinery, which the
MachO.cs file needs.

I considered extracting the Abi enum from the MachO.cs file, but it's used in
several places (including externally, such as binding-tools-for-swift), which
means this would be a rather disruptive change.

I also considered modifying MachO.cs to make it possible to build without the
ErrorHandler machinery, but since we'll need it all anyway eventually, I
decided against this option as well.